+++ api/src/main/kotlin/api/routing/abono/AbonoRouting.kt
@@ -4,7 +4,19 @@ import api.extensions.handle
 import io.ktor.server.routing.*
 
 fun Route.abonoRouting() {
-    get("{id}") {
+    get("/{id}") {
         handle { getAbono() }
+    }
+    post("/subscribe") {
+        handle { subscribeAbono() }
+    }
+    post("/unsubscribe") {
+        handle { unsubscribeAbono() }
+    }
+    post("/subscription") {
+        handle { abonoSubscription() }
+    }
+    post("/subscriptions") {
+        handle { abonoSubscriptions() }
     }
 }