+++ api/src/main/kotlin/api/routing/stops/bus/BusStopsClient.kt
@@ -2,6 +2,7 @@ package api.routing.stops.bus
 
 import api.db.checkStopExists
 import api.extensions.getWrapped
+import api.routing.Response.ResponseJson
 import api.routing.stops.buildStopTimesJson
 import api.utils.Pipeline
 import arrow.core.continuations.either
@@ -18,5 +19,5 @@ suspend fun Pipeline.getBusStopTimesResponse() = either {
     val times = getBusStopTimes(fullStopCode).bind()
     if (times.arrives != null) call.caching = CachingOptions(CacheControl.MaxAge(maxAgeSeconds = 30))
     val statusCode = if (times.arrives == null) HttpStatusCode.ServiceUnavailable else HttpStatusCode.OK
-    api.routing.Response.ResponseJson(buildStopTimesJson(times), statusCode)
+    ResponseJson(buildStopTimesJson(times), statusCode)
 }